refactor: update engine to use new updater api (#1)
* Make work with rust_struct patch

Cleaned up the engine code (including disabling for non-release)
and made work with rust_struct patch.

It doesn't (yet) send along all the information that the updater
wants, but at least some of the plumbing is there.

* Make work with rust_struct patch

Cleaned up the engine code (including disabling for non-release)
and made work with rust_struct patch.

It doesn't (yet) send along all the information that the updater
wants, but at least some of the plumbing is there.

* attempt to expose symbols on Android

* Remove MacOS changes

* Remove more MacOS changes