Add support for shorebird.yaml
This loads shorebird.yaml as a string (from the java side) and passes it to the C++ side of the engine which then passes it along (via a C api) to the updater Rust library for parsing.  This plumbs through support for product_id, base_url and channel and we can now add more to shorebird.yaml without having to touch this code again.