Fix libapp.so path passed to shorebird. (#4)
Previously the path was always "libapp.so", for whatever reason the
engine currently has a quirk of having ["libapp.so", "real/path/libapp.so"]
in that array.  Presumably a bug.

I also fixed shorebird_current_version to shorebird_current_patch_number
while I was editing the engine since there was a TODO in the header.