Teach our fork of the engine to build on Android arm32
This required conditionally linking against separate builds of
libupdater.a as well as providing a stub for getauxval.

If we find a way to move off of using ring (via rust-tls, via reqwests)
we could get rid of this.  More likely Flutter will just move to a
newer version of SDK/NDK:
https://github.com/flutter/flutter/issues/82000