fix: Use vm_data and vm_instrs from the base snapshot
We previously were using vm_data and vm_instrs from the patch
but we've learned there are Stubs which are loaded from
vm_data, and are not wrapped.  We looked at wrapping them, but
certain ones like the intitial "execute dart code" stub can't
easily be wrapped (the Dart THR register isn't set yet).

This isn't well tested and mixed mode is blocked on further changes
but I think it works?